fix imports after main restructure
