+++ src/main/java/com/group4/app/model/actions/MoveAction.java
@@ -4,6 +4,7 @@ import java.util.List;
 import java.util.Set;
 
 import com.group4.app.model.Model;
+import com.group4.app.model.creatures.ICreatureManager;
 import com.group4.app.model.dungeon.IPositionable;
 import com.group4.app.model.dungeon.PathfindingHelper;
 import com.group4.app.model.dungeon.Position;
