+++ src/main/java/com/group4/app/model/creatures/ICreatureManager.java
@@ -1,9 +1,11 @@
 package com.group4.app.model.creatures;
 
 import com.group4.app.model.IModel;
-import com.group4.app.model.ITurnTaker;
-import com.group4.app.model.Position;
+import com.group4.app.model.dungeon.IPositionable;
 import com.group4.app.model.dungeon.ITileContainer;
+import com.group4.app.model.dungeon.Position;
+import com.group4.app.model.turns.ITurnTaker;
+
 
 public interface ICreatureManager extends IModel {
     void remove(IPositionable e);
