+++ src/main/java/com/group4/app/model/creatures/IPlayerManager.java
@@ -1,6 +1,7 @@
 package com.group4.app.model.creatures;
 
-import com.group4.app.model.actions.ActionInput;
+
+import com.group4.app.model.input.ActionInput;
 
 public interface IPlayerManager extends ICreatureManager {
     void setPlayerDied();
