+++ src/test/java/com/group4/app/testmodel/TestTurnHandler.java
@@ -1,14 +1,11 @@
 package com.group4.app.testmodel;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 
-import com.group4.app.model.Position;
 import org.junit.jupiter.api.Test;
 
-import com.group4.app.model.creatures.EnemyFactory;
 import com.group4.app.model.turns.ITurnTaker;
 import com.group4.app.model.turns.TurnHandler;
